requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
astroid==4.0.3
asttokens==3.0.1
attrs==25.4.0
Authlib==1.6.7
autoflake==2.3.1
backcall==0.2.0
bandit==1.9.3
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
codespell==2.4.1
colorama==0.4.6
coverage==7.13.3
cryptography==46.0.4
cyclonedx-python-lib==11.6.0
dash==4.0.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
docopt==0.6.2
dparse==0.6.4
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
flake8==7.3.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipython==8.12.3
isort==7.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
librt==0.7.8
license-expression==30.4.4
mando==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.2.0
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.4.2
packageurl-python==0.17.6
packaging==26.0
panda==0.3.1
pandas==3.0.0
pandocfilters==1.5.1
parso==0.8.5
pathspec==1.0.4
pexpect==4.9.0
pickleshare==0.7.5
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
pipreqs==0.5.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
py-serializable==2.1.0
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.11.0
pylint==4.0.4
PyNaCl==1.6.2
pyparsing==3.3.2
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
pyzmq==27.1.0
radon==6.0.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
retrying==1.4.2
rich==14.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
stack-data==0.6.3
stevedore==5.6.0
tabulate==0.9.0
tenacity==9.1.3
tinycss2==1.4.0
tomli==2.4.0
tomli_w==1.2.0
tomlkit==0.14.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vulture==2.14
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.5
yarg==0.1.9
zipp==3.23.0
